refactor(cloud): extract gdrive CLI into gdrive_cli module
